Make uploading an optional argument that must be passed.

+++ bin/rsync.pl
@@ -3,8 +3,14 @@
 use strict;
 use warnings;
 
+use Getopt::Long;
+my $upload;
+my $result = GetOptions ("upload" => \$upload);
+die "bad args" unless $result;
+
 chdir 'web' || die $!;
 system ('/usr/local/bin/ttree', '--EVAL_PERL', '-f', 'lib/site.cfg', '-a');
-system('rsync', '--delete', '-qa',  'html/', '/home/justin/web/adelaidepm/');
 
+system('rsync', '--delete', '-qa',  'html/', '/home/justin/web/adelaidepm/')
+  if $upload;
